post_sync interface integrated
