+++ tests/test_post_sync.py
@@ -1,5 +1,5 @@
 import pytest
-from user_sync.post_sync import post_sync
+from user_sync.post_sync import manager
 
 
 @pytest.fixture
@@ -18,15 +18,15 @@ def example_user():
 
 def test_add_umapi_user(example_user, monkeypatch):
     with monkeypatch.context() as m:
-        m.setattr(post_sync, '_SYNC_DATA_STORE', {})
+        m.setattr(manager, '_SYNC_DATA_STORE', {})
         email_id = 'user@example.com'
-        post_sync.update_sync_data(email_id, 'umapi', [], [], **example_user)
-        assert post_sync._SYNC_DATA_STORE[email_id.lower()]['umapi_data'] == example_user
+        manager.update_sync_data(email_id, 'umapi', [], [], **example_user)
+        assert manager._SYNC_DATA_STORE[email_id.lower()]['umapi_data'] == example_user
 
 
 def test_add_directory_user(example_user, monkeypatch):
     with monkeypatch.context() as m:
-        m.setattr(post_sync, '_SYNC_DATA_STORE', {})
+        m.setattr(manager, '_SYNC_DATA_STORE', {})
         email_id = 'user@example.com'
         example_user['raw_data'] = {
             'mail': 'user@example.com',
@@ -36,40 +36,40 @@ def test_add_directory_user(example_user, monkeypatch):
             'memberOf': [],
             'department': 'IT',
         }
-        post_sync.update_sync_data(email_id, 'directory', [], [], **example_user)
-        assert post_sync._SYNC_DATA_STORE[email_id.lower()]['directory_data'] == example_user
+        manager.update_sync_data(email_id, 'directory', [], [], **example_user)
+        assert manager._SYNC_DATA_STORE[email_id.lower()]['directory_data'] == example_user
 
 
 def test_add_groups(example_user, monkeypatch):
     with monkeypatch.context() as m:
-        m.setattr(post_sync, '_SYNC_DATA_STORE', {})
+        m.setattr(manager, '_SYNC_DATA_STORE', {})
         email_id = 'user@example.com'
         example_user['groups'] = ['group1', 'group2', 'group3']
         groups_add = ['group3', 'group4', 'group5']
-        post_sync.update_sync_data(email_id, 'umapi', groups_add, [], **example_user)
-        assert sorted(post_sync._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(list(set(
+        manager.update_sync_data(email_id, 'umapi', groups_add, [], **example_user)
+        assert sorted(manager._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(list(set(
             example_user['groups']) | set(groups_add)))
 
 
 def test_remove_groups(example_user, monkeypatch):
     with monkeypatch.context() as m:
-        m.setattr(post_sync, '_SYNC_DATA_STORE', {})
+        m.setattr(manager, '_SYNC_DATA_STORE', {})
         email_id = 'user@example.com'
         example_user['groups'] = ['group1', 'group2', 'group3']
         groups_remove = ['group1', 'group2']
-        post_sync.update_sync_data(email_id, 'umapi', [], groups_remove, **example_user)
-        assert sorted(post_sync._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(list(set(
+        manager.update_sync_data(email_id, 'umapi', [], groups_remove, **example_user)
+        assert sorted(manager._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(list(set(
             example_user['groups']) - set(groups_remove)))
 
 
 def test_add_remove_groups(example_user, monkeypatch):
     with monkeypatch.context() as m:
-        m.setattr(post_sync, '_SYNC_DATA_STORE', {})
+        m.setattr(manager, '_SYNC_DATA_STORE', {})
         email_id = 'user@example.com'
         example_user['groups'] = ['group1', 'group2', 'group3', 'group4', 'group5']
         groups_add = ['group6']
         groups_remove = ['group1', 'group2']
-        post_sync.update_sync_data(email_id, 'umapi', groups_add, groups_remove, **example_user)
+        manager.update_sync_data(email_id, 'umapi', groups_add, groups_remove, **example_user)
         delta_groups = list(set(example_user['groups']) | set(groups_add))
         delta_groups = list(set(delta_groups) - set(groups_remove))
-        assert sorted(post_sync._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(delta_groups)
+        assert sorted(manager._SYNC_DATA_STORE[email_id.lower()]['umapi_data']['groups']) == sorted(delta_groups)
