+++ user_sync/post_sync/__init__.py
@@ -1,6 +1,10 @@
-def init():
-    pass
+class PostSyncConnector:
+    def __init__(self):
+        pass
+
+    def run(self):
+        pass
 
 
-class PostSyncConnector(object):
-    pass
+
+
